@@ -1,6 +1,6 @@
 ﻿namespace DataFilters.Grammar.Parsing
 {
-    using DataFilters.Grammar.Syntax;
+    using Syntax;
 
     using Superpower;
     using Superpower.Model;
@@ -20,7 +20,7 @@
     public static class FilterTokenParser
     {
         /// <summary>
-        /// Parser for many <see cref="FilterToken.Letter"/>
+        /// Parser for many <see cref="FilterToken.Letter"/>s or <see cref="FilterToken.Digit"/>s.
         /// </summary>
         public static TokenListParser<FilterToken, ConstantValueExpression> AlphaNumeric => from data in (
                                                                                                from symbolBefore in Token.EqualTo(FilterToken.Escaped).Try()
@@ -33,11 +33,11 @@
                                                                                                                         .Or(Token.EqualTo(FilterToken.None).Try())
                                                                                                                         .Or(Token.EqualTo(FilterToken.Dot).Try())
                                                                                                                         .Many()
-                                                                                               where symbolBefore.AtLeastOnce()
-                                                                                                     || digitsBefore.AtLeastOnce()
-                                                                                                     || alpha.AtLeastOnce()
-                                                                                                     || digitsAfter.AtLeastOnce()
-                                                                                                     || symbolAfter.AtLeastOnce()
+                                                                                               where symbolBefore.Length > 0
+                                                                                                     || digitsBefore.Length > 0
+                                                                                                     || alpha.Length > 0
+                                                                                                     || digitsAfter.Length > 0
+                                                                                                     || symbolAfter.Length > 0
 
                                                                                                let value = string.Concat(string.Concat(symbolBefore.Select(x => x.ToStringValue())),
                                                                                                                          string.Concat(digitsBefore.Select(x => x.ToStringValue())),
@@ -81,90 +81,110 @@
         /// <summary>
         /// Parser for "starts with" expressions
         /// </summary>
-        public static TokenListParser<FilterToken, StartsWithExpression> StartsWith => (from data in AlphaNumeric.AtLeastOnce()
+        public static TokenListParser<FilterToken, StartsWithExpression> StartsWith => (from text in Text
                                                                                         from _ in Asterisk
-                                                                                        select new StartsWithExpression(string.Concat(data.Select(x => x.Value)))).Try()
+                                                                                        select new StartsWithExpression(text)).Try()
                                                                                         .Or(
-                                                                                            from text in Text
+                                                                                            from data in AlphaNumeric.AtLeastOnce()
                                                                                             from _ in Asterisk
-                                                                                            select new StartsWithExpression(text)
+                                                                                            select new StartsWithExpression(string.Concat(data.Select(x => x.Value)))
                                                                                         );
 
         /// <summary>
-        /// Parser for "starts with" expressions
+        /// Parser for "ends with" expressions
         /// </summary>
-        public static TokenListParser<FilterToken, EndsWithExpression> EndsWith => Asterisk.IgnoreThen(AlphaNumeric.AtLeastOnce())
-                                                                                           .Select(data => new EndsWithExpression(string.Concat(data.Select(item => item.Value)))).Try()
-                                                                                   .Or(Asterisk.IgnoreThen(Text)
-                                                                                       .Select(text => new EndsWithExpression(text)));
+        public static TokenListParser<FilterToken, EndsWithExpression> EndsWith => Asterisk.IgnoreThen(Text)
+                                                                                           .Select(text => new EndsWithExpression(text)).Try()
+                                                                                   .Or(Asterisk.IgnoreThen(AlphaNumeric.AtLeastOnce())
+                                                                                       .Select(data => new EndsWithExpression(string.Concat(data.Select(item => item.Value)))));
 
         /// <summary>
         /// Parser for "contains" expression
         /// </summary>
-        public static TokenListParser<FilterToken, ContainsExpression> Contains => (from _ in Asterisk
-                                                                                    from data in (
-                                                                                        from symbolBefore in Token.EqualTo(FilterToken.Escaped).Try().Or(Whitespace).Many()
-                                                                                        from puncBefore in Punctuation.Many()
-                                                                                        from alpha in AlphaNumeric.Many()
-                                                                                        from puncAfter in Punctuation.Many()
-                                                                                        from symbolAfter in Token.EqualTo(FilterToken.Escaped).Try().Or(Whitespace).Many()
-                                                                                        where symbolBefore.AtLeastOnce()
-                                                                                              || puncBefore.AtLeastOnce()
-                                                                                              || alpha.AtLeastOnce()
-                                                                                              || puncAfter.AtLeastOnce()
-                                                                                              || symbolAfter.AtLeastOnce()
-                                                                                        select new
-                                                                                        {
-                                                                                            value = string.Concat(string.Concat(symbolBefore.Select(x => x.ToStringValue())),
-                                                                                                                  string.Concat(puncBefore.Select(x => x.Value)),
-                                                                                                                  string.Concat(alpha.Select(item => item.Value)),
-                                                                                                                  string.Concat(puncAfter.Select(x => x.Value)),
-                                                                                                                  string.Concat(symbolAfter.Select(x => x.ToStringValue())))
-                                                                                        }).AtLeastOnce()
+        public static TokenListParser<FilterToken, ContainsExpression> Contains => Text.Between(Asterisk, Asterisk)
+                                                                                       .Select(text => new ContainsExpression(text)).Try()
+                                                                                       .Or(
+                                                                                           from _ in Asterisk
+                                                                                           from data in (
+                                                                                               from beforeSymbol in Token.EqualTo(FilterToken.Escaped).Try().Or(Whitespace).Many()
+                                                                                               from beforePunctuation in Punctuation.Many()
+                                                                                               from alpha in AlphaNumeric.Many()
+                                                                                               from afterPunctuation in Punctuation.Many()
+                                                                                               from afterSymbol in Token.EqualTo(FilterToken.Escaped).Try().Or(Whitespace).Many()
+                                                                                               where beforeSymbol.AtLeastOnce()
+                                                                                                     || beforePunctuation.AtLeastOnce()
+                                                                                                     || alpha.AtLeastOnce()
+                                                                                                     || afterPunctuation.AtLeastOnce()
+                                                                                                     || afterSymbol.AtLeastOnce()
+                                                                                               select new
+                                                                                               {
+                                                                                                   value = string.Concat(string.Concat(beforeSymbol.Select(x => x.ToStringValue())),
+                                                                                                                         string.Concat(beforePunctuation.Select(x => x.Value)),
+                                                                                                                         string.Concat(alpha.Select(item => item.Value)),
+                                                                                                                         string.Concat(afterPunctuation.Select(x => x.Value)),
+                                                                                                                         string.Concat(afterSymbol.Select(x => x.ToStringValue())))
+                                                                                               }).AtLeastOnce()
                                                                                     from escaped in Token.EqualTo(FilterToken.Backslash).Many()
                                                                                     from __ in Asterisk
-                                                                                    select new ContainsExpression(string.Concat(data.Select(x => x.value)))).Try()
-                                                                                   .Or(Text.Between(Asterisk, Asterisk)
-                                                                                            .Select(text => new ContainsExpression(text)));
+                                                                                    select new ContainsExpression(string.Concat(data.Select(x => x.value))));
 
         /// <summary>
         /// Parser for logical OR expression.
         /// </summary>
-        public static TokenListParser<FilterToken, FilterExpression> Or => Parse.Chain(Token.EqualTo(FilterToken.Or),
-                                                                                       UnaryExpression,
-                                                                                       (_, left, right) => new OrExpression(left, right));
+        public static TokenListParser<FilterToken, OrExpression> Or
+            => from left in Parse.Ref(() => UnaryExpression)
+               from _ in Token.EqualTo(FilterToken.Or)
+               from right in Parse.Ref(() => UnaryExpression)
+               select left | right;
 
         /// <summary>
         /// Parser for logical AND expression
         /// </summary>
-        public static TokenListParser<FilterToken, AndExpression> And => (from left in UnaryExpression
+        public static TokenListParser<FilterToken, AndExpression> And => (from left in Parse.Ref(() => UnaryExpression)
                                                                           from _ in Token.EqualTo(FilterToken.And)
-                                                                          from right in UnaryExpression
+                                                                          from right in Parse.Ref(() => UnaryExpression)
                                                                           select new AndExpression(left, right)).Try()
                                                                         .Or(from left in AlphaNumeric
                                                                             from _ in Asterisk
                                                                             from right in AlphaNumeric
+                                                                            from __ in Asterisk
                                                                             select new AndExpression(new StartsWithExpression(left.Value),
-                                                                                                    new EndsWithExpression(right.Value)));
+                                                                                                    new ContainsExpression(right.Value))).Try()
+                                                                        .Or(from left in AlphaNumeric
+                                                                            from _ in Asterisk
+                                                                            from right in AlphaNumeric
+                                                                            select new AndExpression(new StartsWithExpression(left.Value),
+                                                                                                    new EndsWithExpression(right.Value)))
+                                                                        ;
 
         /// <summary>
         /// Parser for NOT expression
         /// </summary>
-        public static TokenListParser<FilterToken, NotExpression> Not => from _ in Token.EqualTo(FilterToken.Bang)
+        public static TokenListParser<FilterToken, NotExpression> Not => (from bangs in Token.EqualTo(FilterToken.Bang).AtLeastOnce()
                                                                          from expression in Parse.Ref(() => UnaryExpression)
-                                                                         select new NotExpression(expression);
+                                                                         select (bangs, expression))
+            .Select(tuple =>
+            {
+                NotExpression notExpression = !tuple.expression;
+
+                for (int i = 1; i < tuple.bangs.Length; i++)
+                {
+                    notExpression = !notExpression;
+                }
+                return notExpression;
+            });
 
         private static TokenListParser<FilterToken, BracketExpression> Bracket => (
-                                                                                    from _ in Token.EqualTo(FilterToken.OpenSquaredBracket)
+                                                                                    from _ in Token.EqualTo(FilterToken.LeftSquaredBracket)
                                                                                     from rangeValues in BuildRangeBracketValuesParser(Alpha).Or(BuildRangeBracketValuesParser(Digit)).AtLeastOnce()
-                                                                                    from __ in Token.EqualTo(FilterToken.CloseSquaredBracket)
+                                                                                    from __ in Token.EqualTo(FilterToken.RightSquaredBracket)
                                                                                     select new BracketExpression(rangeValues)
                                                                                 ).Try()
                                                                                 .Or
                                                                                 (
-                                                                                    from _ in Token.EqualTo(FilterToken.OpenSquaredBracket)
+                                                                                    from _ in Token.EqualTo(FilterToken.LeftSquaredBracket)
                                                                                     from alpha in Alpha.Try().Or(Digit).AtLeastOnce()
-                                                                                    from __ in Token.EqualTo(FilterToken.CloseSquaredBracket)
+                                                                                    from __ in Token.EqualTo(FilterToken.RightSquaredBracket)
                                                                                     select new BracketExpression(new ConstantBracketValue(alpha.Select(chr => chr.ToStringValue())
                                                                                                                                                      .Aggregate((total, current) => $"{total}{current}"))
                                                                                                                 )
@@ -172,7 +192,7 @@
         /// <summary>
         /// Builds a parser that can extract bracket expression values
         /// </summary>
-        /// <param name="token">The parser that can parse values inside a <c>[</c> qnd <c>]</c></param>
+        /// <param name="token">The parser that can parse values inside a <c>[</c> and <c>]</c></param>
         /// <returns></returns>
         private static TokenListParser<FilterToken, BracketValue> BuildRangeBracketValuesParser(TokenListParser<FilterToken, Token<FilterToken>> token)
             => (from rangeStart in token
@@ -193,11 +213,11 @@
             {
                 return  // Case [ min TO max ] 
                         (
-                            from _ in Token.EqualTo(FilterToken.OpenSquaredBracket)
+                            from _ in Token.EqualTo(FilterToken.LeftSquaredBracket)
                             from min in Constant
                             from __ in RangeSeparator
                             from max in Constant
-                            from ___ in Token.EqualTo(FilterToken.CloseSquaredBracket)
+                            from ___ in Token.EqualTo(FilterToken.RightSquaredBracket)
 
                             where min != default || max != default
                             select new IntervalExpression(
@@ -225,12 +245,12 @@
                         ).Try()
                       // Case ] min TO max ] : lower bound excluded from the interval
                       .Or(
-                            from _ in Token.EqualTo(FilterToken.CloseSquaredBracket)
+                            from _ in Token.EqualTo(FilterToken.RightSquaredBracket)
                             from min in Asterisk.Try().Cast<FilterToken, AsteriskExpression, FilterExpression>()
                                                 .Or(Constant)
                             from __ in RangeSeparator
                             from max in Constant
-                            from ___ in Token.EqualTo(FilterToken.CloseSquaredBracket)
+                            from ___ in Token.EqualTo(FilterToken.RightSquaredBracket)
 
                             where min != default || max != default
                             select new IntervalExpression(
@@ -259,13 +279,13 @@
                         ).Try()
                       // Case syntax [ min TO max [ : upper bound excluded from the interval
                       .Or(
-                            from _ in Token.EqualTo(FilterToken.OpenSquaredBracket)
+                            from _ in Token.EqualTo(FilterToken.LeftSquaredBracket)
                             from min in Constant
                             from __ in RangeSeparator
                             from max in Asterisk.Try().Cast<FilterToken, AsteriskExpression, FilterExpression>()
                                                 .Or(Constant)
 
-                            from ___ in Token.EqualTo(FilterToken.OpenSquaredBracket)
+                            from ___ in Token.EqualTo(FilterToken.LeftSquaredBracket)
 
                             where min != default || max != default
                             select new IntervalExpression(
@@ -294,14 +314,14 @@
                         ).Try()
                       // Case  ] min TO max [ : lower and upper bounds excluded from the interval
                       .Or(
-                            from _ in Token.EqualTo(FilterToken.CloseSquaredBracket)
+                            from _ in Token.EqualTo(FilterToken.RightSquaredBracket)
                             from min in Asterisk.Try().Cast<FilterToken, AsteriskExpression, FilterExpression>()
                                                 .Or(Constant)
 
                             from __ in RangeSeparator
                             from max in Asterisk.Try().Cast<FilterToken, AsteriskExpression, FilterExpression>()
                                                 .Or(Constant)
-                            from ___ in Token.EqualTo(FilterToken.OpenSquaredBracket)
+                            from ___ in Token.EqualTo(FilterToken.LeftSquaredBracket)
 
                             where min != default || max != default
                             select new IntervalExpression(
@@ -332,16 +352,15 @@
             }
         }
 
-        private static TokenListParser<FilterToken, FilterExpression> Constant => GlobalUniqueIdentifier.Try().Cast<FilterToken, GuidValueExpression, FilterExpression>()
-                                                                                                          .Or(DateAndTime.Try().Cast<FilterToken, DateTimeExpression, FilterExpression>())
-                                                                                                          .Or(Time.Try().Cast<FilterToken, TimeExpression, FilterExpression>())
-                                                                                                          .Or(Date.Try().Cast<FilterToken, DateExpression, FilterExpression>())
-                                                                                                          .Or(Duration.Try().Cast<FilterToken, DurationExpression, FilterExpression>())
-                                                                                                          .Or(Number.Try().Cast<FilterToken, NumericValueExpression, FilterExpression>())
-                                                                                                          .Or(Bool.Try().Cast<FilterToken, StringValueExpression, FilterExpression>())
-                                                                                                          .Or(Text.Cast<FilterToken, TextExpression, FilterExpression>())
-                                                                                                          .Or(AlphaNumeric.Cast<FilterToken, ConstantValueExpression, FilterExpression>())
-            ;
+        private static TokenListParser<FilterToken, FilterExpression> Constant => Parse.OneOf(GlobalUniqueIdentifier.Try().Cast<FilterToken, GuidValueExpression, FilterExpression>(),
+                                                                                              DateAndTime.Try().Cast<FilterToken, DateTimeExpression, FilterExpression>(),
+                                                                                              Time.Try().Cast<FilterToken, TimeExpression, FilterExpression>(),
+                                                                                              Date.Try().Cast<FilterToken, DateExpression, FilterExpression>(),
+                                                                                              Duration.Try().Cast<FilterToken, DurationExpression, FilterExpression>(),
+                                                                                              Number.Try().Cast<FilterToken, NumericValueExpression, FilterExpression>(),
+                                                                                              Bool.Try().Cast<FilterToken, StringValueExpression, FilterExpression>(),
+                                                                                              Text.Cast<FilterToken, TextExpression, FilterExpression>(),
+                                                                                              AlphaNumeric.Cast<FilterToken, ConstantValueExpression, FilterExpression>());
 
         private static TokenListParser<FilterToken, Token<FilterToken>> RangeSeparator => from _ in Whitespace.AtLeastOnce()
                                                                                           from rangeSeparator in Token.EqualToValue(FilterToken.Letter, "T")
@@ -354,25 +373,24 @@
         /// <summary>
         /// Group expression
         /// </summary>
-        public static TokenListParser<FilterToken, GroupExpression> Group => from _ in Token.EqualTo(FilterToken.OpenParenthese)
-                                                                             from expression in BinaryOrUnaryExpression
-                                                                             from __ in Token.EqualTo(FilterToken.CloseParenthese)
+        public static TokenListParser<FilterToken, GroupExpression> Group => from _ in Token.EqualTo(FilterToken.LeftParenthesis)
+                                                                             from expression in Parse.Ref(() => BinaryOrUnaryExpression)
+                                                                             from __ in Token.EqualTo(FilterToken.RightParenthesis)
                                                                              select new GroupExpression(expression);
 
-        private static TokenListParser<FilterToken, FilterExpression> BinaryOrUnaryExpression => Parse.Ref(() => Not.Try().Cast<FilterToken, NotExpression, FilterExpression>())
-                                                                                                      .Or(Parse.Ref(() => And.Try().Cast<FilterToken, AndExpression, FilterExpression>()))
-                                                                                                      .Or(Parse.Ref(() => Or.Try()))
-                                                                                                      .Or(Parse.Ref(() => OneOf.Try().Cast<FilterToken, OneOfExpression, FilterExpression>()))
-                                                                                                      .Or(Parse.Ref(() => UnaryExpression));
+        private static TokenListParser<FilterToken, FilterExpression> BinaryOrUnaryExpression => Parse.OneOf(Parse.Ref(() => And.Try().Cast<FilterToken, AndExpression, FilterExpression>()),
+                                                                                                             Parse.Ref(() => Or.Try().Cast<FilterToken, OrExpression, FilterExpression>()),
+                                                                                                             Parse.Ref(() => UnaryExpression ))
+        ;
 
         /// <summary>
         /// Property name parser
         /// </summary>
         public static TokenListParser<FilterToken, PropertyName> Property => from prop in AlphaNumeric
                                                                              from subProps in (
-                                                                                 from _ in Token.EqualTo(FilterToken.OpenSquaredBracket)
+                                                                                 from _ in Token.EqualTo(FilterToken.LeftSquaredBracket)
                                                                                  from subProp in AlphaNumeric.Between(Token.EqualTo(FilterToken.DoubleQuote), Token.EqualTo(FilterToken.DoubleQuote))
-                                                                                 from __ in Token.EqualTo(FilterToken.CloseSquaredBracket)
+                                                                                 from __ in Token.EqualTo(FilterToken.RightSquaredBracket)
                                                                                  select @$"[""{subProp.Value}""]"
                                                                              ).Many()
                                                                              select new PropertyName(string.Concat(prop.Value, string.Concat(subProps)));
@@ -398,15 +416,16 @@
                     value switch
                     {
                         RangeBracketValue rangeValue => Enumerable.Range(rangeValue.Start, rangeValue.End - rangeValue.Start + 1)
-                                                                    .Select(ascii => (char)ascii),
-                        ConstantBracketValue constantValue => constantValue.Value.ToCharArray(),
+                                                                  .Select(ascii => (char)ascii),
+                        ConstantBracketValue constantValue => [.. constantValue.Value],
 #if NET7_0_OR_GREATER
                         _ => throw new UnreachableException("Unexpected regex value")
 #else
                         _ => throw new NotSupportedException("Unexpected regex value")
 #endif
                     })
-                .SelectMany(x => x);
+                // Flatten collections
+                .SelectMany(chr => chr);
 
         /// <summary>
         /// Parser for expressions that contains one or more regex parts.
@@ -418,7 +437,6 @@
                                         .Or(AlphaNumeric.Cast<FilterToken, ConstantValueExpression, FilterExpression>())
                                         .OptionalOrDefault()
                 from tail in Bracket
-
                 select (head: (FilterExpression)head, (object)body, tail: (FilterExpression)tail)
                 ).Try()
                 .Or(
@@ -437,42 +455,18 @@
                     select (head, body: (object)body, tail)
                 ).Try()
                 .Or(
-                    from head in Group
-                    from body in Asterisk.Try().Cast<FilterToken, AsteriskExpression, FilterExpression>()
-                                            .Or(AlphaNumeric.Cast<FilterToken, ConstantValueExpression, FilterExpression>())
-                                            .OptionalOrDefault()
-                    from tail in Group
-
-                    select (head: (FilterExpression)head, (object)body, tail: (FilterExpression)tail)
-                ).Try()
-                .Or(
                     from head in EndsWith.Try().Cast<FilterToken, EndsWithExpression, FilterExpression>()
                                             .Or(StartsWith.Try().Cast<FilterToken, StartsWithExpression, FilterExpression>())
-                                            .Or(AlphaNumeric.Cast<FilterToken, ConstantValueExpression, FilterExpression>())
-                                            .Or(Asterisk.Cast<FilterToken, AsteriskExpression, FilterExpression>())
+                                            .Or(Constant.Try())
+                                            .Or(Asterisk.Try().Cast<FilterToken, AsteriskExpression, FilterExpression>())
                                             .OptionalOrDefault()
-                    from body in Group
-                    from tail in EndsWith.Try().Cast<FilterToken, EndsWithExpression, FilterExpression>()
-                                        .Or(StartsWith.Try().Cast<FilterToken, StartsWithExpression, FilterExpression>())
-                                        .Or(AlphaNumeric.Try().Cast<FilterToken, ConstantValueExpression, FilterExpression>())
-                                        .Or(Asterisk.Try().Cast<FilterToken, AsteriskExpression, FilterExpression>())
-                                        .OptionalOrDefault()
-
-                    select (head, body: (object)body, tail)
-                ).Try()
-                .Or(
-                    from head in EndsWith.Try().Cast<FilterToken, EndsWithExpression, FilterExpression>()
-                                            .Or(StartsWith.Try().Cast<FilterToken, StartsWithExpression, FilterExpression>())
-                                            .Or(AlphaNumeric.Cast<FilterToken, ConstantValueExpression, FilterExpression>())
-                                            .Or(Asterisk.Cast<FilterToken, AsteriskExpression, FilterExpression>())
-                                            .OptionalOrDefault()
-                    from _ in Token.EqualTo(FilterToken.LeftBrace)
+                    from _ in Token.EqualTo(FilterToken.LeftCurlyBrace)
                     from body in Constant.AtLeastOnceDelimitedBy(Token.EqualTo(FilterToken.Or))
-                    from __ in Token.EqualTo(FilterToken.RightBrace)
+                    from __ in Token.EqualTo(FilterToken.RightCurlyBrace)
                     from tail in EndsWith.Try().Cast<FilterToken, EndsWithExpression, FilterExpression>()
                                             .Or(StartsWith.Try().Cast<FilterToken, StartsWithExpression, FilterExpression>())
-                                            .Or(AlphaNumeric.Cast<FilterToken, ConstantValueExpression, FilterExpression>())
-                                            .Or(Asterisk.Cast<FilterToken, AsteriskExpression, FilterExpression>())
+                                            .Or(Constant.Try())
+                                            .Or(Asterisk.Try().Cast<FilterToken, AsteriskExpression, FilterExpression>())
                                             .OptionalOrDefault()
 
                     select (head, body: (object)body.Select(item => item).OfType<ConstantValueExpression>().ToArray(), tail)
@@ -481,67 +475,66 @@
             {
                 return item switch
                 {
-                    // *<regex>
-                    (AsteriskExpression, BracketExpression bracket, null) => new OneOfExpression(ConvertRegexToCharArray(bracket.Values).Select(chr => new EndsWithExpression(chr.ToString()))
-                                                                                                                     .ToArray()),
-                    // *<regex><constant>
-                    (AsteriskExpression, BracketExpression bracket, ConstantValueExpression tail) => new OneOfExpression(ConvertRegexToCharArray(bracket.Values).Select(chr => new EndsWithExpression($"{chr}{tail.Value}"))
-                                                                                                                                             .ToArray()),
+                    // *<bracket>
+                    (AsteriskExpression, BracketExpression bracket, null) => new OneOfExpression([ .. ConvertRegexToCharArray(bracket.Values).Select(chr => new EndsWithExpression(chr.ToString())) ]),
 
-                    // <regex>*
-                    (null, BracketExpression bracket, AsteriskExpression) => new OneOfExpression(ConvertRegexToCharArray(bracket.Values).Select(chr => new StartsWithExpression(chr.ToString()))
-                                                                                                                     .ToArray()),
+                    // *<bracket><constant>
+                    (AsteriskExpression, BracketExpression bracket, ConstantValueExpression tail) => new OneOfExpression(ConvertRegexToCharArray(bracket.Values).Select(FilterExpression (chr) => new EndsWithExpression($"{chr}{tail.Value}")).ToArray()),
 
-                    // <endswith><regex>
-                    (EndsWithExpression head, BracketExpression body, null) => new OneOfExpression(ConvertRegexToCharArray(body.Values).Select(chr => new EndsWithExpression($"{head.Value}{chr}"))
-                                                                                                                      .ToArray()),
-                    // <endswith><regex><constant>
-                    (EndsWithExpression head, BracketExpression body, ConstantValueExpression constant) => new OneOfExpression(ConvertRegexToCharArray(body.Values).Select(chr => new EndsWithExpression($"{head.Value}{chr}{constant.Value}"))
-                                                                                                                      .ToArray()),
-                    // <startswith><regex>
-                    (StartsWithExpression head, BracketExpression body, null) => new OneOfExpression(ConvertRegexToCharArray(body.Values).Select(chr => new AndExpression(head,
-                                                                                                                                                                           new EndsWithExpression(chr.ToString())))
-                                                                                                                  .ToArray()),
-                    // <regex><startwith>*
-                    (null, BracketExpression bracket, StartsWithExpression body) => new OneOfExpression(ConvertRegexToCharArray(bracket.Values).Select(chr => new StartsWithExpression($"{chr}{body.Value}")).ToArray()),
+                    // <bracket>*
+                    (null, BracketExpression bracket, AsteriskExpression) => new OneOfExpression([.. ConvertRegexToCharArray(bracket.Values).Select(chr => new StartsWithExpression(chr.ToString()))]),
 
-                    // <constant><regex><constant>
-                    (ConstantValueExpression bracket, BracketExpression regex, ConstantValueExpression tail) => new OneOfExpression(ConvertRegexToCharArray(regex.Values).Select(chr => new StringValueExpression($"{bracket.Value}{chr}{tail.Value}")).ToArray()),
-                    // <constant><regex>
-                    (ConstantValueExpression head, BracketExpression bracket, null) => new OneOfExpression(ConvertRegexToCharArray(bracket.Values).Select(chr => new StringValueExpression($"{head.Value}{chr}")).ToArray()),
+                    // <ends with><bracket>
+                    (EndsWithExpression head, BracketExpression body, null) => new OneOfExpression(ConvertRegexToCharArray(body.Values).Select(chr => (FilterExpression) new EndsWithExpression($"{head.Value}{chr}")).ToArray()),
 
-                    // <regex><constant>
-                    (null, BracketExpression bracket, ConstantValueExpression tail) => new OneOfExpression(ConvertRegexToCharArray(bracket.Values).Select(chr => new StringValueExpression($"{chr}{tail.Value}")).ToArray()),
+                    // <ends with><bracket><constant>
+                    (EndsWithExpression head, BracketExpression body, ConstantValueExpression constant) => new OneOfExpression([.. ConvertRegexToCharArray(body.Values).Select(chr => new EndsWithExpression($"{head.Value}{chr}{constant.Value}"))]),
 
-                    // <regex>
-                    (null, BracketExpression bracket, null) => new OneOfExpression(ConvertRegexToCharArray(bracket.Values).Select(chr => new StringValueExpression(chr.ToString())).ToArray()),
+                    // <starts with><bracket>
+                    (StartsWithExpression head, BracketExpression body, null) => new OneOfExpression([.. ConvertRegexToCharArray(body.Values).Select(chr => new AndExpression(head, new EndsWithExpression(chr.ToString())))]),
 
-                    // <regex><constant><regex>
-                    (BracketExpression head, ConstantValueExpression body, BracketExpression tail) => new(ConvertRegexToCharArray(head.Values).CrossJoin(ConvertRegexToCharArray(tail.Values))
+                    // <bracket><starts with>*
+                    (null, BracketExpression bracket, StartsWithExpression body) => new OneOfExpression([.. ConvertRegexToCharArray(bracket.Values).Select(chr => new StartsWithExpression($"{chr}{body.Value}"))]),
+
+                    // <constant><bracket><constant>
+                    (ConstantValueExpression bracket, BracketExpression regex, ConstantValueExpression tail) => new OneOfExpression([.. ConvertRegexToCharArray(regex.Values).Select(chr => new StringValueExpression($"{bracket.Value}{chr}{tail.Value}"))]),
+
+                    // <constant><bracket>
+                    (ConstantValueExpression head, BracketExpression bracket, null) => new OneOfExpression([..ConvertRegexToCharArray(bracket.Values).Select(chr => new StringValueExpression($"{head.Value}{chr}"))]),
+
+                    // <bracket><constant>
+                    (null, BracketExpression bracket, ConstantValueExpression tail) => new OneOfExpression([.. ConvertRegexToCharArray(bracket.Values).Select(chr => new StringValueExpression($"{chr}{tail.Value}"))]),
+
+                    // <bracket>
+                    (null, BracketExpression bracket, null) => new OneOfExpression([..ConvertRegexToCharArray(bracket.Values).Select(chr => new StringValueExpression(chr.ToString()))]),
+
+                    // <bracket><constant><bracket>
+                    (BracketExpression head, ConstantValueExpression body, BracketExpression tail) => new OneOfExpression([ ..ConvertRegexToCharArray(head.Values).CrossJoin(ConvertRegexToCharArray(tail.Values))
                                                                             .Select(tuple => (start: tuple.Item1, end: tuple.Item2))
                                                                             .Select(tuple => new StringValueExpression($"{tuple.start}{body.Value}{tuple.end}"))
-                                                                            .ToArray()),
-                    // <regex><regex>
-                    (BracketExpression head, null, BracketExpression tail) => new(ConvertRegexToCharArray(head.Values).CrossJoin(ConvertRegexToCharArray(tail.Values))
+                                                                            ]),
+                    // <bracket><bracket>
+                    (BracketExpression head, null, BracketExpression tail) => new OneOfExpression([..ConvertRegexToCharArray(head.Values).CrossJoin(ConvertRegexToCharArray(tail.Values))
                                                                             .Select(tuple => (start: tuple.Item1, end: tuple.Item2))
                                                                             .Select(tuple => new StringValueExpression($"{tuple.start}{tuple.end}"))
-                                                                            .ToArray()),
-                    // <regex><endswith>
-                    (null, BracketExpression body, EndsWithExpression tail) => new OneOfExpression(ConvertRegexToCharArray(body.Values).Select(chr => new AndExpression(new StartsWithExpression(chr.ToString()), tail)).ToArray()),
+                                                                            ]),
+                    // <bracket><ends with>
+                    (null, BracketExpression body, EndsWithExpression tail) => new OneOfExpression([.. ConvertRegexToCharArray(body.Values).Select(chr => new AndExpression(new StartsWithExpression(chr.ToString()), tail))]),
 
-                    // *<oneof>
-                    (AsteriskExpression, IEnumerable<ConstantValueExpression> constants, null) => new OneOfExpression(constants.Select(constant => new EndsWithExpression(constant.Value)).ToArray()),
+                    // *<one of>
+                    (AsteriskExpression, IEnumerable<ConstantValueExpression> constants, null) => new OneOfExpression([.. constants.Select(constant => new EndsWithExpression(constant.Value))]),
 
-                    // <oneof>*
-                    (null, IEnumerable<ConstantValueExpression> constants, AsteriskExpression) => new OneOfExpression(constants.Select(constant => new StartsWithExpression(constant.Value)).ToArray()),
+                    // <one of>*
+                    (null, IEnumerable<ConstantValueExpression> constants, AsteriskExpression) => new OneOfExpression([.. constants.Select(constant => new StartsWithExpression(constant.Value))]),
 
-                    // *<oneof>*
-                    (AsteriskExpression, IEnumerable<ConstantValueExpression> constants, AsteriskExpression) => new OneOfExpression(constants.Select(constant => new ContainsExpression(constant.Value)).ToArray()),
-                    // <oneof><endswith>
-                    (null, IEnumerable<ConstantValueExpression> constants, EndsWithExpression endsWith) => new OneOfExpression(constants.Select(constant => constant + endsWith).ToArray()),
-                    // <oneof>
-                    (null, IEnumerable<ConstantValueExpression> constants, null) => new OneOfExpression(constants.ToArray()),
+                    // *<one of>*
+                    (AsteriskExpression, IEnumerable<ConstantValueExpression> constants, AsteriskExpression) => new OneOfExpression([.. constants.Select(constant => new ContainsExpression(constant.Value))]),
 
+                    // <one of><ends with>
+                    (null, IEnumerable<ConstantValueExpression> constants, EndsWithExpression endsWith) => new OneOfExpression([..constants.Select(constant => constant + endsWith)]),
+
+                    // <one of>
+                    (null, IEnumerable<ConstantValueExpression> constants, null) => new OneOfExpression([.. constants]),
 #if NET7_0_OR_GREATER
                     _ => throw new UnreachableException($"Unsupported {nameof(OneOf)} expression :  {item}")
 #else
@@ -553,16 +546,17 @@
         /// <summary>
         /// Parser for Date and Time
         /// </summary>
+        /// <returns>a <see cref="DateTimeExpression"/></returns>
         public static TokenListParser<FilterToken, DateTimeExpression> DateAndTime => (from date in Date
                                                                                        from separator in Token.EqualToValue(FilterToken.Letter, "T")
                                                                                                               .Or(Token.EqualTo(FilterToken.Whitespace))
                                                                                        from time in Time
                                                                                        from offset in Offset.OptionalOrDefault()
                                                                                        select new DateTimeExpression(date,
-                                                                                                                     new(hours: time.Hours,
-                                                                                                                         minutes: time.Minutes,
-                                                                                                                         seconds: time.Seconds,
-                                                                                                                         milliseconds: time.Milliseconds),
+                                                                                                                     new TimeExpression(hours: time.Hours,
+                                                                                                                                        minutes: time.Minutes,
+                                                                                                                                        seconds: time.Seconds,
+                                                                                                                                        milliseconds: time.Milliseconds),
                                                                                                                      offset)).Try()
                                                                                       // DATE
                                                                                       .Or(from date in Date
@@ -576,7 +570,7 @@
         private static TokenListParser<FilterToken, OffsetExpression> Offset => (from _ in Token.EqualToValue(FilterToken.Letter, "Z")
                                                                                  select OffsetExpression.Zero).Try()
                                                                           .Or(
-                                                                                from sign in MinusOrPlusSign.OptionalOrDefault(NumericSign.Plus)
+                                                                                from sign in MinusOrPlusSign.OptionalOrDefault()
                                                                                 from hourDigits in IntDigits(2)
                                                                                 from _ in Colon
                                                                                 from minuteDigits in IntDigits(2)
@@ -638,27 +632,41 @@
                                                                                                          int.Parse(TokensToString(secondDigits), CultureInfo.InvariantCulture)));
 
         /// <summary>
-        /// Parses all supported expressions
+        /// Parses all supported unary expressions
         /// </summary>
-        private static TokenListParser<FilterToken, FilterExpression> UnaryExpression => Parse.Ref(() => Not.Try().Cast<FilterToken, NotExpression, FilterExpression>())
-                                                                                              .Or(Parse.Ref(() => Group.Try().Cast<FilterToken, GroupExpression, FilterExpression>()))
-                                                                                              .Or(Parse.Ref(() => OneOf.Try().Cast<FilterToken, OneOfExpression, FilterExpression>()))
-                                                                                              .Or(Parse.Ref(() => Interval.Try().Cast<FilterToken, IntervalExpression, FilterExpression>()))
-                                                                                              .Or(Parse.Ref(() => GlobalUniqueIdentifier.Try().Cast<FilterToken, GuidValueExpression, FilterExpression>()))
-                                                                                              .Or(Parse.Ref(() => Duration.Try().Cast<FilterToken, DurationExpression, FilterExpression>()))
-                                                                                              .Or(Parse.Ref(() => DateAndTime.Try().Cast<FilterToken, DateTimeExpression, FilterExpression>()))
-                                                                                              .Or(Parse.Ref(() => Time.Try().Cast<FilterToken, TimeExpression, FilterExpression>()))
-                                                                                              .Or(Parse.Ref(() => Date.Try().Cast<FilterToken, DateExpression, FilterExpression>()))
-                                                                                              .Or(Parse.Ref(() => Contains.Try().Cast<FilterToken, ContainsExpression, FilterExpression>()))
-                                                                                              .Or(Parse.Ref(() => StartsWith.Try().Cast<FilterToken, StartsWithExpression, FilterExpression>()))
-                                                                                              .Or(Parse.Ref(() => EndsWith.Try().Cast<FilterToken, EndsWithExpression, FilterExpression>()))
-                                                                                              .Or(Parse.Ref(() => Bool.Try().Cast<FilterToken, StringValueExpression, FilterExpression>()))
-                                                                                              .Or(Parse.Ref(() => Text.Try().Cast<FilterToken, TextExpression, FilterExpression>()))
-                                                                                              .Or(Parse.Ref(() => AlphaNumeric.Try().Cast<FilterToken, ConstantValueExpression, FilterExpression>()))
-                                                                                              .Or(Parse.Ref(() => Number.Cast<FilterToken, NumericValueExpression, FilterExpression>()))
+        private static TokenListParser<FilterToken, FilterExpression> UnaryExpression => Parse.OneOf(Parse.Ref(() => OneOf.Try().Cast<FilterToken, OneOfExpression, FilterExpression>()),
+                                                                                                     Parse.Ref(() => Not.Try().Cast<FilterToken, NotExpression, FilterExpression>()),
+                                                                                                     Parse.Ref(() => Group.Try().Cast<FilterToken, GroupExpression, FilterExpression>()),
+                                                                                                     Parse.Ref(() => Interval.Try().Cast<FilterToken, IntervalExpression, FilterExpression>()),
+                                                                                                     Parse.Ref(() => GlobalUniqueIdentifier.Try().Cast<FilterToken, GuidValueExpression, FilterExpression>()),
+                                                                                                     Parse.Ref(() => Duration.Try().Cast<FilterToken, DurationExpression, FilterExpression>()),
+                                                                                                     Parse.Ref(() => DateAndTime.Try().Cast<FilterToken, DateTimeExpression, FilterExpression>()),
+                                                                                                     Parse.Ref(() => Time.Try().Cast<FilterToken, TimeExpression, FilterExpression>()),
+                                                                                                     Parse.Ref(() => Date.Try().Cast<FilterToken, DateExpression, FilterExpression>()),
+                                                                                              // Special case of <constant>*<constant> which should be turned into an AndExpression
+                                                                                              // even though from a user perspective it's just a text filter with no specific logic in it
+                                                                                                    Parse.Ref(() => StartsAndEndsWith.Try().Cast<FilterToken, AndExpression, FilterExpression>()),
+                                                                                                    Parse.Ref(() => Contains.Try().Cast<FilterToken, ContainsExpression, FilterExpression>()),
+                                                                                                    Parse.Ref(() => EndsWith.Try().Cast<FilterToken, EndsWithExpression, FilterExpression>()),
+                                                                                                    Parse.Ref(() => StartsWith.Try().Cast<FilterToken, StartsWithExpression, FilterExpression>()),
+                                                                                                    Parse.Ref(() => Bool.Try().Cast<FilterToken, StringValueExpression, FilterExpression>()),
+                                                                                                    Parse.Ref(() => Text.Try().Cast<FilterToken, TextExpression, FilterExpression>()),
+                                                                                                    Parse.Ref(() => AlphaNumeric.Try().Cast<FilterToken, ConstantValueExpression, FilterExpression>()),
+                                                                                                    Parse.Ref(() => Number.Cast<FilterToken, NumericValueExpression, FilterExpression>()))
             ;
+
         /// <summary>
-        /// Parser for <c>property=value</c> pair.
+        /// Handles special case of &lt;constant&gt;*&lt;constant&gt; which should be turned into a <see cref="AndExpression"/>
+        /// even though from user's perspective, it's just a text filter with no specific logic in it.
+        /// </summary>
+        private static TokenListParser<FilterToken, AndExpression> StartsAndEndsWith
+            => from alphaBeforeAsterisk in AlphaNumeric
+                from __ in Asterisk
+                from alphaAfterAsterisk in AlphaNumeric
+                select new StartsWithExpression(alphaBeforeAsterisk.Value) & new EndsWithExpression(alphaAfterAsterisk.Value);
+
+        /// <summary>
+        /// Parser for a <c>property=&lt;expression&gt;</c> pair.
         /// </summary>
         public static TokenListParser<FilterToken, (PropertyName, FilterExpression)> Criterion => from property in Property
                                                                                                   from _ in Token.EqualTo(FilterToken.Equal)
@@ -672,11 +680,28 @@
 
         private static TokenListParser<FilterToken, NumericValueExpression> FloatOrDouble
             => (from sign in MinusOrPlusSign.Optional()
-                from digitBeforeDot in Token.EqualTo(FilterToken.Digit).AtLeastOnce()
+                from beforeDotDigits in Token.EqualTo(FilterToken.Digit).AtLeastOnce()
                 from _ in Token.EqualTo(FilterToken.Dot)
-                from digitAfterDot in Token.EqualTo(FilterToken.Digit).AtLeastOnce()
-                from __ in Token.EqualToValueIgnoreCase(FilterToken.Letter, "d")
-                select new NumericValueExpression($"{ConvertSignToChar(sign)}{TokensToString(digitBeforeDot)}.{TokensToString(digitAfterDot)}"))
+                from afterDotDigits in Token.EqualTo(FilterToken.Digit).AtLeastOnce()
+                from __ in Token.EqualToValueIgnoreCase(FilterToken.Letter, "e")
+                from exponentSign in MinusOrPlusSign.Optional()
+                from afterExponentSignDigits in IntegerOrLong
+                select new NumericValueExpression($"{ConvertSignToChar(sign)}{TokensToString(beforeDotDigits)}.{TokensToString(afterDotDigits)}E{ConvertSignToChar(exponentSign, true)}{afterExponentSignDigits.EscapedParseableString}"))
+                .Try()
+                .Or(
+                    from sign in MinusOrPlusSign.Optional()
+                    from beforeDotDigits in Token.EqualTo(FilterToken.Digit).AtLeastOnce()
+                    from __ in Token.EqualToValueIgnoreCase(FilterToken.Letter, "e")
+                    from exponentSign in MinusOrPlusSign.Optional()
+                    from afterExponentSignDigits in IntegerOrLong
+                    select new NumericValueExpression($"{ConvertSignToChar(sign)}{TokensToString(beforeDotDigits)}E{ConvertSignToChar(exponentSign, true)}{afterExponentSignDigits.EscapedParseableString}"))
+                .Try()
+                .Or(from sign in MinusOrPlusSign.Optional()
+                    from digitBeforeDot in Token.EqualTo(FilterToken.Digit).AtLeastOnce()
+                    from _ in Token.EqualTo(FilterToken.Dot)
+                    from digitAfterDot in Token.EqualTo(FilterToken.Digit).AtLeastOnce()
+                    from __ in Token.EqualToValueIgnoreCase(FilterToken.Letter, "d")
+                    select new NumericValueExpression($"{ConvertSignToChar(sign)}{TokensToString(digitBeforeDot)}.{TokensToString(digitAfterDot)}"))
                 .Try()
                 .Or(
                      from sign in MinusOrPlusSign.Optional()
@@ -706,6 +731,11 @@
                 from digits in Token.EqualTo(FilterToken.Digit).AtLeastOnce()
                 from hint in Token.EqualToValueIgnoreCase(FilterToken.Letter, "L")
                 select new NumericValueExpression($"{ConvertSignToChar(sign)}{TokensToString(digits)}"))
+        .Try()
+            .Or(from sign in MinusOrPlusSign.Optional()
+                from digits in Token.EqualTo(FilterToken.Digit).AtLeastOnce()
+                from hint in Token.EqualToValueIgnoreCase(FilterToken.Letter, "L")
+                select new NumericValueExpression($"{ConvertSignToChar(sign)}{TokensToString(digits)}"))
                 .Try()
                 .Or(
                     from sign in MinusOrPlusSign.Optional()
@@ -714,17 +744,19 @@
                 )
             ;
 
-        private static string ConvertSignToChar(NumericSign? sign) => sign switch
-        {
-            NumericSign.Minus => "-",
-            _ => string.Empty,
-        };
+        private static string ConvertSignToChar(NumericSign? sign, bool mustOutputSign = false)
+            => sign switch
+            {
+                NumericSign.Minus => "-",
+                _ => mustOutputSign ? "+" : string.Empty
+            };
 
         /// <summary>
         /// Parser for many <see cref="Criterion"/> separated by <c>&amp;</c>.
         /// </summary>
-        public static TokenListParser<FilterToken, (PropertyName, FilterExpression)[]> Criteria => from criteria in Criterion.ManyDelimitedBy(Token.EqualTo(FilterToken.Ampersand))
-                                                                                                   select criteria;
+        public static TokenListParser<FilterToken, (PropertyName, FilterExpression)[]> Criteria
+            => from criteria in Criterion.ManyDelimitedBy(Token.EqualTo(FilterToken.Ampersand))
+               select criteria;
 
         private static TokenListParser<FilterToken, StringValueExpression> Punctuation =>
         (
@@ -741,43 +773,44 @@
             select new StringValueExpression(c.ToStringValue())
         );
 
-        private static TokenListParser<FilterToken, GuidValueExpression> GlobalUniqueIdentifier => from chr1 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
-                                                                                                   from chr2 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
-                                                                                                   from chr3 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
-                                                                                                   from chr4 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
-                                                                                                   from chr5 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
-                                                                                                   from chr6 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
-                                                                                                   from chr7 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
-                                                                                                   from chr8 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
-                                                                                                   from _ in Token.EqualTo(FilterToken.Dash)
-                                                                                                   from chr9 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
-                                                                                                   from chr10 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
-                                                                                                   from chr11 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
-                                                                                                   from chr12 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
-                                                                                                   from __ in Token.EqualTo(FilterToken.Dash)
-                                                                                                   from chr13 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
-                                                                                                   from chr14 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
-                                                                                                   from chr15 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
-                                                                                                   from chr16 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
-                                                                                                   from ___ in Token.EqualTo(FilterToken.Dash)
-                                                                                                   from chr17 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
-                                                                                                   from chr18 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
-                                                                                                   from chr19 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
-                                                                                                   from chr20 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
-                                                                                                   from ____ in Token.EqualTo(FilterToken.Dash)
-                                                                                                   from chr21 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
-                                                                                                   from chr22 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
-                                                                                                   from chr23 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
-                                                                                                   from chr24 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
-                                                                                                   from chr25 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
-                                                                                                   from chr26 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
-                                                                                                   from chr27 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
-                                                                                                   from chr28 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
-                                                                                                   from chr29 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
-                                                                                                   from chr30 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
-                                                                                                   from chr31 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
-                                                                                                   from chr32 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
-                                                                                                   select new GuidValueExpression($"{TokensToString(new[] { chr1, chr2, chr3, chr4, chr5, chr6, chr7, chr8 })}-{TokensToString(new[] { chr9, chr10, chr11, chr12 })}-{TokensToString(new[] { chr13, chr14, chr15, chr16 })}-{TokensToString(new[] { chr17, chr18, chr19, chr20 })}-{TokensToString(new[] { chr21, chr22, chr23, chr24, chr25, chr26, chr27, chr28, chr29, chr30, chr31, chr32 })}");
+        private static TokenListParser<FilterToken, GuidValueExpression> GlobalUniqueIdentifier
+            => from chr1 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
+               from chr2 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
+               from chr3 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
+               from chr4 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
+               from chr5 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
+               from chr6 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
+               from chr7 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
+               from chr8 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
+               from _ in Token.EqualTo(FilterToken.Dash)
+               from chr9 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
+               from chr10 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
+               from chr11 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
+               from chr12 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
+               from __ in Token.EqualTo(FilterToken.Dash)
+               from chr13 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
+               from chr14 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
+               from chr15 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
+               from chr16 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
+               from ___ in Token.EqualTo(FilterToken.Dash)
+               from chr17 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
+               from chr18 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
+               from chr19 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
+               from chr20 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
+               from ____ in Token.EqualTo(FilterToken.Dash)
+               from chr21 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
+               from chr22 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
+               from chr23 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
+               from chr24 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
+               from chr25 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
+               from chr26 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
+               from chr27 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
+               from chr28 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
+               from chr29 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
+               from chr30 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
+               from chr31 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
+               from chr32 in Token.EqualTo(FilterToken.Letter).Try().Or(Token.EqualTo(FilterToken.Digit))
+               select new GuidValueExpression($"{TokensToString([chr1, chr2, chr3, chr4, chr5, chr6, chr7, chr8 ])}-{TokensToString([chr9, chr10, chr11, chr12 ])}-{TokensToString([chr13, chr14, chr15, chr16 ])}-{TokensToString([chr17, chr18, chr19, chr20 ])}-{TokensToString([chr21, chr22, chr23, chr24, chr25, chr26, chr27, chr28, chr29, chr30, chr31, chr32 ])}");
 
         /// <summary>
         /// Parser for duration
@@ -817,9 +850,9 @@
 
         private static string TokensToString(IEnumerable<Token<FilterToken>> tokens)
         {
-            static string TokenToString(Token<FilterToken> token) => token.ToStringValue();
-
             return string.Concat(tokens.Select(TokenToString));
+
+            static string TokenToString(Token<FilterToken> token) => token.ToStringValue();
         }
     }
 }
