@@ -1,14 +1,14 @@
-﻿namespace DataFilters.Grammar.Syntax
+﻿using System.Linq.Expressions;
+
+namespace DataFilters.Grammar.Syntax
 {
     using System;
     using System.Collections.Generic;
     using System.Linq;
     using System.Text;
-#if !NETSTANDARD1_3
-    using Ardalis.GuardClauses;
-#endif
 
-    using static DataFilters.Grammar.Parsing.FilterTokenizer;
+
+    using static Parsing.FilterTokenizer;
 
     /// <summary>
     /// A <see cref="FilterExpression"/> that holds a string value
@@ -26,44 +26,40 @@
         /// Builds a new <see cref="ContainsExpression"/> that holds the specified <paramref name="value"/>.
         /// </summary>
         /// <param name="value"></param>
-        /// <exception cref="ArgumentNullException">if <paramref name="value"/> is <c>null</c></exception>
+        /// <exception cref="ArgumentNullException">if <paramref name="value"/> is <see langword="null"/></exception>
         /// <exception cref="ArgumentOutOfRangeException">if <paramref name="value"/> is <c>empty</c></exception>
         public ContainsExpression(string value)
         {
-            if (value is null)
+            Value = value switch
             {
-                throw new ArgumentNullException(nameof(value));
-            }
+                null => throw new ArgumentNullException(nameof(value)),
+                {Length: 0} => throw new ArgumentOutOfRangeException(nameof(value)),
+                _ => value
+            };
 
-            if (value.Length == 0)
+            _lazyEscapedParseableString = new Lazy<string>(() =>
             {
-                throw new ArgumentOutOfRangeException(nameof(value), $"{nameof(value)} cannot be empty");
-            }
-
-            Value = value;
-
-            _lazyEscapedParseableString = new(() =>
-            {
-                // The length of the final parseable string in worst case scenarios will double (1 backlash for each character of the original input)
-                // Also we need an extra position for the final '*' that will be appended in all cases
+                // The length of the final parseable string in worst case scenarios will double (1 backlash for each special character from the original input)
+                // Also we need two extra positions for '*' that will be prepended and appended in all cases
                 bool requireEscapingCharacters = Value.AtLeastOnce(chr => SpecialCharacters.Contains(chr));
                 StringBuilder parseableString;
 
                 if (requireEscapingCharacters)
                 {
-                    parseableString = new((Value.Length * 2) + 2);
+                    parseableString = new StringBuilder(( Value.Length * 2 ) + 2);
                     foreach (char chr in Value)
                     {
                         if (SpecialCharacters.Contains(chr))
                         {
                             parseableString = parseableString.Append('\\');
                         }
+
                         parseableString = parseableString.Append(chr);
                     }
                 }
                 else
                 {
-                    parseableString = new(Value, Value.Length + 2);
+                    parseableString = new StringBuilder(Value, Value.Length + 2);
                 }
 
                 return parseableString.Insert(0, "*").Append('*').ToString();
@@ -71,25 +67,33 @@
         }
 
         /// <summary>
-        /// Builds a new <see cref="StartsWithExpression"/> that holds the specified <paramref name="text"/>.
+        /// Builds a new <see cref="ContainsExpression"/> that holds the specified <paramref name="text"/>.
         /// </summary>
         /// <param name="text"></param>
-        /// <exception cref="ArgumentNullException"><paramref name="text"/> is <c>null</c>.</exception>
+        /// <exception cref="ArgumentNullException"><paramref name="text"/> is <see langword="null"/>.</exception>
         public ContainsExpression(TextExpression text)
-#if !NETSTANDARD1_3
-            : this (Guard.Against.Null(text, nameof(text)).OriginalString)
-        {
-            _lazyEscapedParseableString = new(() => $"*{text.EscapedParseableString}*");
-        }
-#else
-        {
-            if (text is null)
+            : this(text switch
             {
-                throw new ArgumentNullException(nameof(text));
-            }
-            _lazyEscapedParseableString = new(() => $"*{text.EscapedParseableString}*");
+                null => throw new ArgumentNullException(nameof(text)),
+                _ =>  text.OriginalString,
+            })
+        {
+            _lazyEscapedParseableString = new Lazy<string>(() =>
+            {
+                StringBuilder sb = new StringBuilder((text.OriginalString.Length *  2) + 2);
+
+                foreach (char chr in text.OriginalString)
+                {
+                    if (chr is '\\' or '\"')
+                    {
+                        sb = sb.Append('\\');
+                    }
+                    sb.Append(chr);
+                }
+
+                return sb.Insert(0, "*\"").Append('\"').Append('*').ToString();
+            });
         }
-#endif
 
         ///<inheritdoc/>
         public override double Complexity => 1.5;
@@ -105,5 +109,15 @@
 
         ///<inheritdoc/>
         public override string EscapedParseableString => _lazyEscapedParseableString.Value;
+
+        ///<inheritdoc/>
+        public override bool IsEquivalentTo(FilterExpression other)
+            => other switch
+            {
+                ContainsExpression contains => Equals(contains),
+                GroupExpression { Expression: var innerExpression } => innerExpression.IsEquivalentTo(this),
+                ISimplifiable simplifiable => simplifiable.Simplify().IsEquivalentTo(this),
+                _ => false
+            };
     }
 }
