@@ -5,7 +5,7 @@ namespace DataFilters.Grammar.Syntax
     /// <summary>
     /// A <see cref="FilterExpression"/> implementation that contains values associated to a duration (see "https://en.wikipedia.org/wiki/ISO_8601#Durations")
     /// </summary>
-    public sealed class DurationExpression : FilterExpression, IEquatable<DurationExpression>
+    public sealed class DurationExpression : FilterExpression, IEquatable<DurationExpression>, IFormattable
     {
         /// <summary>
         /// Years part of the expression
