@@ -9,18 +9,19 @@
     /// <summary>
     /// a <see cref="FilterExpression"/> that contains multiple <see cref="FilterExpression"/>s as <see cref="Values"/>.
     /// </summary>
-    public sealed class OneOfExpression : FilterExpression, IEquatable<OneOfExpression>, ISimplifiable
+    public sealed class OneOfExpression : FilterExpression, IEquatable<OneOfExpression>, ISimplifiable, IEnumerable<FilterExpression>
     {
         private static readonly ArrayEqualityComparer<FilterExpression> EqualityComparer = new();
 
         /// <summary>
         /// Collection of <see cref="FilterExpression"/> that the current instance holds.
         /// </summary>
-        public IReadOnlyCollection<FilterExpression> Values => _values.ToList().AsReadOnly();
+        public IReadOnlyList<FilterExpression> Values => [ .. _values];
 
         private readonly FilterExpression[] _values;
 
-        private readonly Lazy<string> _lazyParseableString;
+        private readonly Lazy<string> _lazyEscapedParseableString;
+        private readonly Lazy<string> _lazyOriginalString;
 
         /// <summary>
         /// Builds a new <see cref="OneOfExpression"/> instance.
@@ -40,10 +41,10 @@
                 throw new InvalidOperationException($"{nameof(OneOfExpression)} cannot be empty");
             }
 
-            _values = values.Where(x => x is not null)
-                            .ToArray();
+            _values = [.. values.Where(x => x is not null)];
 
-            _lazyParseableString = new(() => $"{{{string.Join(",", Values.Select(v => v.EscapedParseableString))}}}");
+            _lazyEscapedParseableString = new Lazy<string>(() => $"{{{string.Join("|", Values.Select(v => v.EscapedParseableString))}}}");
+            _lazyOriginalString = new Lazy<string>(() => $"{{{string.Join("|", Values.Select(v => v.OriginalString))}}}");
         }
 
         /// <inheritdoc/>
@@ -61,11 +62,12 @@
                                     || !(_values.Except(oneOfExpression._values).Any() || oneOfExpression._values.Except(_values).Any())
                     : other switch
                     {
-                        AsteriskExpression asterisk => Simplify().IsEquivalentTo(asterisk),
-                        ConstantValueExpression constant => Simplify().IsEquivalentTo(constant),
-                        DateExpression date => Simplify().IsEquivalentTo(date),
-                        DateTimeExpression dateTime => Simplify().IsEquivalentTo(dateTime),
-                        ISimplifiable simplifiable => Simplify().IsEquivalentTo(simplifiable.Simplify()),
+                        AsteriskExpression asterisk => Values.All(x => x is AsteriskExpression),
+                        ConstantValueExpression constant => Values.All(value => value.IsEquivalentTo(constant)),
+                        DateExpression date => Values.All(value => value.IsEquivalentTo(date)),
+                        DateTimeExpression dateTime => Values.All(value => value.Equals(dateTime) || value.IsEquivalentTo(dateTime)),
+                        OrExpression or => Values.All(value => value.Equals(or.Left) || value.Equals(or.Right) || value.IsEquivalentTo(or.Left) || value.IsEquivalentTo(or.Right)),
+                        ISimplifiable simplifiable => Values.All(value => value.IsEquivalentTo(simplifiable.Simplify())),
                         _ => false
                     };
             }
@@ -95,32 +97,18 @@
         public FilterExpression Simplify()
         {
             HashSet<FilterExpression> curatedExpressions = [];
-
-            foreach (IGrouping<bool, FilterExpression> expr in _values.ToLookup(x => x is OneOfExpression))
+            
+            foreach (FilterExpression expression in Values)
             {
-                if (expr.Key)
-                {
-                    OneOfExpression oneOfExpression = new(expr.Select(item => ((OneOfExpression)item).Values)
-                                                              .SelectMany(x => x)
-                                                              .ToArray());
+                FilterExpression simplified = expression.As<ISimplifiable>()?.Simplify() ?? expression;
 
-                    curatedExpressions.Add(oneOfExpression.Simplify());
-                }
-                else
-                {
-                    foreach (FilterExpression expression in expr)
-                    {
-                        FilterExpression simplifiedExpression = (expression as ISimplifiable)?.Simplify() ?? expression;
-                        if (!curatedExpressions.Contains(simplifiedExpression) && !curatedExpressions.Any(existing => existing.IsEquivalentTo(simplifiedExpression)))
-                        {
-                            curatedExpressions.Add(simplifiedExpression);
-                        }
-                    }
-                }
+                curatedExpressions.RemoveWhere(val => val.Equals(simplified) || (val.IsEquivalentTo(simplified) 
+                                                      && val.Complexity > simplified.Complexity)); 
+                curatedExpressions.Add(simplified);
+
             }
 
             FilterExpression simplifiedResult;
-
             switch (curatedExpressions.Count)
             {
                 case 1:
@@ -129,9 +117,7 @@
                 case 2:
                     FilterExpression first = curatedExpressions.First();
                     FilterExpression other = curatedExpressions.Last();
-                    simplifiedResult = first is OneOfExpression oneOfFirst && other is OneOfExpression oneOfSecond
-                        ? new OneOfExpression([.. oneOfFirst.Values, .. oneOfSecond.Values])
-                        : new OrExpression(first, other);
+                    simplifiedResult = new OrExpression(first, other);
                     break;
                 default:
                     simplifiedResult = new OneOfExpression([.. curatedExpressions]);
@@ -142,6 +128,33 @@
         }
 
         ///<inheritdoc/>
-        public override string EscapedParseableString => _lazyParseableString.Value;
+        public override string EscapedParseableString => _lazyEscapedParseableString.Value;
+
+        /// <inheritdoc />
+        public override string ToString() => _lazyOriginalString.Value;
+
+        /// <inheritdoc />
+        IEnumerator<FilterExpression> IEnumerable<FilterExpression>.GetEnumerator()
+        {
+            foreach (FilterExpression expression in _values)
+            {
+                yield return expression;
+            }
+        }
+
+        /// <inheritdoc />
+        System.Collections.IEnumerator System.Collections.IEnumerable.GetEnumerator() => _values.GetEnumerator();
+
+        /// <inheritdoc />
+        public override string ToString(string format, IFormatProvider formatProvider)
+        {
+            FormattableString formattable = format switch
+            {
+                "D" or "d" => $"{{{string.Join(", ", Values.Select(expression => $"{expression:d}"))}}}",
+                _ => $"{_lazyOriginalString.Value}"
+            };
+
+            return formattable.ToString(formatProvider);
+        }
     }
 }
