@@ -30,8 +30,10 @@
     ///     Value = "John"
     /// };
     ///
+    ///
     /// // Build an expression tree
     /// Expression&lt;Func&lt;Person, bool&gt;&gt; expression = filter.ToExpression&lt;Person&gt;();
+    ///
     ///
     /// // Use the expression to filter data
     /// var filteredData = data.Where(expression.Compile());
@@ -271,10 +273,19 @@
             };
         }
 
+#if NET6_0_OR_GREATER
         /// <summary>
         /// Computes and returns a <see cref="ConstantExpression"/> based on the property expression target type and value.
-        /// Handles different scenarios based on the member type, including <see cref="DateTime"/>, <see cref="DateOnly"/> (for .NET 6.0 or greater), enumerable types, and other types.
+        /// Handles different scenarios based on the member type, including <see cref="DateTime"/>,
+        /// <see cref="DateOnly"/> (for .NET 6.0 or greater),
+        /// enumerable types, and other types.
         /// </summary>
+#else
+        /// <summary>
+        /// Computes and returns a <see cref="ConstantExpression"/> based on the property expression target type and value.
+        /// Handles different scenarios based on the member type, including <see cref="DateTime"/>, enumerable and other types.
+        /// </summary>
+#endif
         private static ConstantExpression ComputeConstantExpressionBasedOnPropertyExpressionTargetTypeAndValue(Type memberType, object value)
         {
             ConstantExpression ce;
@@ -310,7 +321,7 @@
         private static Expression ComputeIsEmpty(MemberExpression property) => IsNotAStringAndIsEnumerable(property.Type)
                             ? Not(Call(typeof(Enumerable),
                                   nameof(Enumerable.Any),
-                                  new Type[] { property.Type.GenericTypeArguments[0] },
+                                  [property.Type.GenericTypeArguments[0]],
                                   property))
                             : Equal(property, Constant(string.Empty));
 
