@@ -1,13 +1,12 @@
-﻿namespace DataFilters.PerfomanceTests
+﻿using DataFilters.PerfomanceTests;
+
+namespace DataFilters.PerformanceTests
 {
     using System;
     using BenchmarkDotNet.Attributes;
     using BenchmarkDotNet.Jobs;
 
     [MemoryDiagnoser]
-    [SimpleJob(RuntimeMoniker.CoreRt31)]
-    [SimpleJob(RuntimeMoniker.Net50)]
-    [SimpleJob(RuntimeMoniker.Net60)]
     [System.Diagnostics.CodeAnalysis.SuppressMessage("Performance", "CA1822:Marquer les membres comme étant static")]
     public class BracketVsOr
     {
