@@ -22,7 +22,7 @@
         public void Ctor_Throws_ArgumentNullException_When_Argument_Is_Null()
         {
             // Act
-            Action action = () => new GroupExpression(null);
+            Action action = () => _ = new GroupExpression(null);
 
             // Assert
             action.Should()
@@ -45,7 +45,7 @@
             => group.Item.Equals(group.Item).Should().BeTrue();
 
         [Property(Arbitrary = [typeof(ExpressionsGenerators)])]
-        public void Equals_should_be_symetric(NonNull<GroupExpression> group, NonNull<FilterExpression> otherExpression)
+        public void Equals_should_be_symmetric(NonNull<GroupExpression> group, NonNull<FilterExpression> otherExpression)
             => group.Item.Equals(otherExpression.Item).Should().Be(otherExpression.Item.Equals(group.Item));
 
         [Property(Arbitrary = [typeof(ExpressionsGenerators)])]
@@ -124,6 +124,12 @@
                     new GroupExpression(new DateTimeExpression(new(2090, 10, 10), new (03,00,40, 583), OffsetExpression.Zero)),
                     true,
                     "Two instances with inner expressions that are equal"
+                },
+                {
+                    new GroupExpression(new NumericValueExpression("0")),
+                    new GroupExpression(new StringValueExpression("0")),
+                    true,
+                    "Two instances with inner expressions that are equal"
                 }
             };
 
@@ -171,7 +177,7 @@
             => group.IsEquivalentTo(group).Should().BeTrue();
 
         [Property(Arbitrary = [typeof(ExpressionsGenerators)])]
-        public void IsEquivalent_should_be_symetric(GroupExpression group, NonNull<FilterExpression> other)
+        public void IsEquivalent_should_be_symmetric(GroupExpression group, NonNull<FilterExpression> other)
         {
             outputHelper.WriteLine($"Group : {group}");
             outputHelper.WriteLine($"Other : {other}");
@@ -179,7 +185,7 @@
             group.IsEquivalentTo(other.Item).Should().Be(other.Item.IsEquivalentTo(group));
         }
 
-        [Property(Arbitrary = [typeof(ExpressionsGenerators)]/*, Replay = "(1500570200792655435,5263687413201141279)"*/)]
+        [Property(Arbitrary = [typeof(ExpressionsGenerators)])]
         public void IsEquivalent_should_be_transitive(GroupExpression group, NonNull<FilterExpression> other, NonNull<FilterExpression> third)
         {
             // Arrange
@@ -200,17 +206,42 @@
             }
         }
 
-        [Property(Arbitrary = [typeof(ExpressionsGenerators)])]
-        public void Given_a_non_null_filter_expression_When_wrapped_inside_a_group_expression_Should_not_change_its_meaning(NonNull<FilterExpression> filterExpression, PositiveInt count)
+        [Property(Arbitrary = [typeof(ExpressionsGenerators)], Replay = "(13485810522331674394,9508666507375819251")]
+        public void Given_a_non_null_filter_expression_When_wrapped_inside_a_group_expression_Should_not_change_its_meaning(NonNull<FilterExpression> filterExpressionGenerator, PositiveInt count)
         {
             // Arrange
             int depth = count.Item / 2;
-            GroupExpression initialGroup = new(filterExpression.Item);
-            GroupExpression otherGroup = new(filterExpression.Item);
+            FilterExpression filterExpression = filterExpressionGenerator.Item;
+            GroupExpression initialGroup = new(filterExpression);
+            GroupExpression otherGroup = new(filterExpression);
 
             for (int i = 0; i < depth; i++)
             {
-                otherGroup = new(otherGroup);
+                otherGroup = new GroupExpression(otherGroup);
+            }
+
+            // Act
+            bool isEquivalent = initialGroup.IsEquivalentTo(otherGroup);
+
+            // Assert
+            outputHelper.WriteLine($"{nameof(initialGroup)}: {initialGroup:d}");
+            outputHelper.WriteLine($"{nameof(otherGroup)}: {otherGroup:d}");
+            isEquivalent.Should().BeTrue();
+        }
+
+        [Property(Arbitrary = [typeof(ExpressionsGenerators)])]
+        public void Given_a_two_groups_that_wraps_the_same_expression_Then_they_should_be_equal(NonNull<FilterExpression> filterExpressionGenerator, PositiveInt count)
+        {
+            // Arrange
+            int depth = count.Item / 2;
+            FilterExpression filterExpression = filterExpressionGenerator.Item;
+            GroupExpression initialGroup = new(filterExpression);
+            GroupExpression otherGroup = new(filterExpression);
+
+            for (int i = 0; i < depth; i++)
+            {
+                initialGroup = new GroupExpression(initialGroup);
+                otherGroup = new GroupExpression(otherGroup);
             }
 
             // Act
@@ -218,6 +249,67 @@
 
             // Assert
             isEquivalent.Should().BeTrue();
+        }
+
+        [Property(Arbitrary = [typeof(ExpressionsGenerators)])]
+        public void Given_a_non_null_FilterExpression_that_is_wrapped_inside_a_GroupExpression_instance_When_calling_Simplify_Then_the_resulting_FilterExpression_should_be_equivalent_to_the_starting_FilterExpression(NonNull<FilterExpression> filterExpressionGenerator)
+        {
+            // Arrange
+            FilterExpression expected = filterExpressionGenerator.Item;
+            GroupExpression group = new(expected);
+
+            // Act
+            FilterExpression actual = group.Simplify();
+
+            // Assert
+            actual.Should().NotBeOfType<GroupExpression>();
+            actual.IsEquivalentTo(expected).Should().BeTrue();
+        }
+
+        public static TheoryData<GroupExpression, FilterExpression> SimplifyCases
+            => new TheoryData<GroupExpression, FilterExpression>()
+            {
+                {
+                    new GroupExpression(new StringValueExpression("prop")),
+                    new StringValueExpression("prop")
+                },
+                {
+                    new GroupExpression(new OrExpression(new StringValueExpression("prop"), new StringValueExpression("prop"))),
+                    new StringValueExpression("prop")
+                }
+            };
+
+        [Theory]
+        [MemberData(nameof(SimplifyCases))]
+        public void Given_GroupExpression_When_calling_Simplify_Then_should_return_expected_outcome(GroupExpression input, FilterExpression expected)
+        {
+            // Act
+            FilterExpression actual = input.Simplify();
+            
+            // Assert
+            actual.Should().Be(expected);
+        }
+
+        public static TheoryData<GroupExpression, FilterExpression, bool, string> IsEquivalentToCases
+            => new()
+            {
+                {
+                    new GroupExpression(new IntervalExpression(new BoundaryExpression(new NumericValueExpression("-1"), true), new BoundaryExpression(new NumericValueExpression("-1"), true))),
+                    new GroupExpression(new IntervalExpression(new BoundaryExpression(new NumericValueExpression("-1"), true), new BoundaryExpression(new NumericValueExpression("-1"), true))),
+                    true,
+                    "left and right are expressions with exactly same values"
+                }
+            };
+        
+        [Theory]
+        [MemberData(nameof(IsEquivalentToCases))]
+        public void Given_left_is_a_GroupExpression_and_right_is_an_expression_Then_IsEquivalent_should_returns_expected_result(GroupExpression left, FilterExpression right, bool expected, string reason)
+        {
+            // Act
+            bool actual = left.IsEquivalentTo(right);
+ 
+            // Assert
+            actual.Should().Be(expected, reason);
         }
     }
 }
